+++ R/change_functions.R
@@ -1,11 +1,11 @@
 #' Calculates the Area of a Raster.
 #'
-#' \code{getArea} reports the area of a RasterLayer object using the pixel
-#' counting method, or the area of a SpatialPolygons object using rgeos::gArea
+#' `getArea` reports the area of a RasterLayer object using the pixel
+#'  counting method, or terra::expanse for SpatRaster and SpatVector objects,
+#'  or the area of a SpatialPolygons or sf object using sf::st_area
 #' @param x Either a RasterLayer or SpatialPolygons object. For a RasterLayer,
 #'   no data value should be NA
-#' @param value.to.count Optional. Value of the cells in a RasterLayer to be
-#'   counted
+#' @param ... Addition arguments based on input format
 #' @return The total area of the cells of interest in km2
 #' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
 #'   \email{calvinkflee@@gmail.com}
@@ -17,56 +17,132 @@
 #' a.r1 <- getArea(r1) # area of all non-NA cells in r1
 #' @export
 #' @import raster
+#' @import terra
+#' @import sf
 
-getArea <- function(x, value.to.count){
+getArea <- function(x, ...){
   if(isLonLat(x)){
     stop('Input raster has a longitude/latitude CRS.\nPlease reproject to a projected coordinate system')
   }
-  if(class(x) == 'RasterLayer'){
-    if(length(raster::unique(x)) != 1 & missing(value.to.count)){
-      warning("The input raster is not binary, counting ALL non NA cells\n")
-      cell.res <- res(x)
-      cell.area <- cell.res[1] * cell.res[2]
-      x.df <- plyr::count(values(x))
-      n.cell <- sum(x.df[which(!is.na(x.df[, 1])), ]$freq)
-      aream2 <- cell.area * n.cell
-      areakm2 <- aream2/1000000
-      return (areakm2)
-    }
-    else if(length(raster::unique(x)) != 1){
-      cell.res <- res(x)
-      cell.area <- cell.res[1] * cell.res[2]
-      x.df <- plyr::count(values(x))
-      n.cell <- x.df[which(x.df[, 1] == value.to.count), ]$freq
-      aream2 <- cell.area * n.cell
-      areakm2 <- aream2/1000000
-      return (areakm2)
-    }
-    else{
-      cell.res <- res(x)
-      cell.area <- cell.res[1] * cell.res[2]
-      x.df <- plyr::count(values(x))
-      n.cell <- x.df[which(!is.na(x.df[,1])), ]$freq
-      aream2 <- cell.area * n.cell
-      areakm2 <- aream2/1000000
-      return (areakm2)
-    }
-  } else if(class(x) == 'SpatialPolygons'){
-    areakm2 <- rgeos::gArea(x) / 1000000
-    return(areakm2)
-  } else stop('Input must be either RasterLayer or SpatialPolygons')
+  UseMethod("getArea", x)
+}
+
+#' Calculates the Area of a Raster from RasterLayer.
+#'
+#' `getArea` reports the area of a RasterLayer object using the pixel
+#'  counting method.
+#' @param x Either a RasterLayer object. No data value should be NA
+#' @param value.to.count Optional. Value of the cells in a RasterLayer to be
+#'   counted
+#' @param ... Addition arguments based on input format
+#' @return The total area of the cells of interest in km2
+#' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
+#'   \email{calvinkflee@@gmail.com}
+#' @family Change functions
+#' @export
+getArea.RasterLayer <- function(x, value.to.count, ...){
+  if(length(raster::unique(x)) != 1 & missing(value.to.count)){
+    warning("The input raster is not binary, counting ALL non NA cells\n")
+    cell.res <- res(x)
+    cell.area <- cell.res[1] * cell.res[2]
+    x.df <- plyr::count(values(x))
+    n.cell <- sum(x.df[which(!is.na(x.df[, 1])), ]$freq)
+    aream2 <- cell.area * n.cell
+    areakm2 <- aream2/1000000
+    return (areakm2)
+  }
+  else if(length(raster::unique(x)) != 1){
+    cell.res <- res(x)
+    cell.area <- cell.res[1] * cell.res[2]
+    x.df <- plyr::count(values(x))
+    n.cell <- x.df[which(x.df[, 1] == value.to.count), ]$freq
+    aream2 <- cell.area * n.cell
+    areakm2 <- aream2/1000000
+    return (areakm2)
+  }
+  else{
+    cell.res <- res(x)
+    cell.area <- cell.res[1] * cell.res[2]
+    x.df <- plyr::count(values(x))
+    n.cell <- x.df[which(!is.na(x.df[,1])), ]$freq
+    aream2 <- cell.area * n.cell
+    areakm2 <- aream2/1000000
+    return (areakm2)
+  }
+}
+
+#' Calculates the Area of a Raster from SpatVect.
+#'
+#' `getArea` reports the area of a SpatVect. object using terra::expanse
+#' @param x A SpatVect object
+#' @param ... Addition arguments based on input format
+#' @return The total area of the cells of interest in km2
+#' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
+#'   \email{calvinkflee@@gmail.com}
+#' @family Change functions
+#' @export
+getArea.SpatVect <- function(x, ...){
+  area <- expanse(x, "km")
+  return(area)
+}
+
+#' Calculates the Area of a Raster from SpatRaster.
+#'
+#' `getArea` reports the area of a SpatRaster object using terra::expanse
+#' @param x SpatRaster
+#' @param byValue Logical. If TRUE, the area for each unique cell value is
+#'    returned.
+#' @param ... Addition arguments based on input format
+#' @return The total area of the cells of interest in km2
+#' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
+#'   \email{calvinkflee@@gmail.com}
+#' @family Change functions
+#' @export
+getArea.SpatRaster <- function(x, byValue, ...){
+  area <- expanse(x, "km", byValue)
+  return(area)
+}
+
+#' Calculates the Area of a Raster from SpatialPolygons.
+#'
+#' `getArea` reports the area of a SpatialPolygons object using sf::st_area
+#' @param x A SpatialPolygons object.
+#' @param ... Addition arguments based on input format
+#' @return The total area of the cells of interest in km2
+#' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
+#'   \email{calvinkflee@@gmail.com}
+#' @family Change functions
+#' @export
+getArea.SpatialPolygons <- function(x, ...){
+  sf_polygon <- st_as_sf(x)
+  area <- st_area(sf_polygon)
+  return(area)
+}
+
+#' Calculates the Area of a Raster from sf object
+#'
+#' `getArea` reports the area of a sf object using sf::st_area
+#' @param x A sf object
+#' @param ... Addition arguments based on input format
+#' @return The total area of the cells of interest in km2
+#' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
+#'   \email{calvinkflee@@gmail.com}
+#' @family Change functions
+#' @export
+getArea.sf <- function(x, ...){
+  area <- st_area(x) / 1000000
+  return(as.numeric(area))
 }
 
 #' Area change between two inputs in km2
 #'
-#' \code{getAreaLoss} reports the difference in area between two inputs. These
-#' can be RasterLayers, SpatialPolygons, or numbers. Any combinations of these
-#' inputs are valid. If using number as input, ensure it is measured in km2
+#' `getAreaLoss` reports the difference in area between two inputs. These
+#' can be RasterLayers, SpatialPolygons, SpatRaster, SpatVect, sf or numbers.
+#' Any combinations of these inputs are valid. If using number as input, ensure
+#' it is measured in km2
 #'
-#' @param x RasterLayer or SpatialPolygons object of distribution or Numeric
-#'   representing area in km2
-#' @param y RasterLayer or SpatialPolygons object of distribution or Numeric
-#'   representing area in km2
+#' @param x Spatial obect or numeric representing area in km2
+#' @param y Spatial object or numeric representing area in km2
 #' @return Returns the difference in area of the two inputs in km2
 #' @author Nicholas Murray \email{murr.nick@@gmail.com}, Calvin Lee
 #'   \email{calvinkflee@@gmail.com}
@@ -81,30 +157,35 @@ getArea <- function(x, value.to.count){
 #' @export
 
 getAreaLoss <- function(x, y){
-  if((class(x) == 'RasterLayer') | (class(x) == 'SpatialPolygons')){
+  if(inherits(x, 'RasterLayer') | inherits(x, 'SpatialPolygons') |
+     inherits(x, 'SpatRaster') | inherits(x, 'SpatVect') | inherits(x, 'sf')){
     a.x <- getArea(x)
   } else if (is.numeric((x))){
     a.x <- x
   } else {
-    stop('x is not a RasterLayer, SpatialPolygons, or Numeric')
+    stop('x is not a RasterLayer, SpatialPolygons, SpatRaster, SpatVect, sf,
+         or Numeric')
   }
-  if((class(y) == 'RasterLayer') | (class(y) == 'SpatialPolygons')){
+  if(inherits(y, 'RasterLayer') | inherits(y, 'SpatialPolygons') |
+     inherits(y, 'SpatRaster') | inherits(y, 'SpatVect') | inherits(y, 'sf')){
     a.y <- getArea(y)
   } else if (is.numeric((y))){
     a.y <- y
   } else {
-    stop('y is not a RasterLayer, SpatialPolygons, or Numeric')
+    stop('y is not a RasterLayer, SpatialPolygons, SpatRaster, SpatVect, sf,
+         or Numeric')
   }
   a.dif.km2 <- (a.x - a.y)
   return(a.dif.km2)
 }
 
+
 #' Change statistics.
 #'
-#' \code{getDeclineStats} calculates the Proportional Rate of Decline (PRD),
+#' `getDeclineStats` calculates the Proportional Rate of Decline (PRD),
 #' Absolute Rate of Decline (ARD) and Annual Rate of Change (ARC), given two
 #' areas at two points in time. Also provides the total area difference. Inputs
-#' are usually the results from \code{getArea}.
+#' are usually the results from `getArea`.
 #'
 #' @param A.t1 Area at time t1
 #' @param year.t1 Year of time t1
@@ -127,7 +208,7 @@ getAreaLoss <- function(x, y){
 #'   Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
 #'   List of Ecosystems Categories and Criteria, Version 1.0. Gland,
 #'   Switzerland: IUCN. ix + 94pp. Available at the following web site:
-#'   \url{https://iucnrle.org/}
+#'   <https://iucnrle.org/>
 #'   Puyravaud, J.-P. 2003. Standardizing the calculation of the
 #'   annual rate of deforestation. Forest Ecology and Management, 177, 593-596.
 #' @examples
@@ -163,8 +244,8 @@ getDeclineStats <- function (A.t1, A.t2, year.t1, year.t2,
 
 #' Future Area Estimate
 #'
-#' \code{futureAreaEstimate} is now deprecated, please use
-#' \code{extrapolateEstimate} instead
+#' `futureAreaEstimate` is now deprecated, please use
+#' `extrapolateEstimate` instead
 #'
 #' @param A.t1 Area at time t1
 #' @param year.t1 Year of time t1
@@ -185,7 +266,7 @@ getDeclineStats <- function (A.t1, A.t2, year.t1, year.t2,
 #'   Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
 #'   List of Ecosystems Categories and Criteria, Version 1.0. Gland,
 #'   Switzerland: IUCN. ix + 94pp. Available at the following web site:
-#'   \url{https://iucnrle.org/}
+#'   <https://iucnrle.org/>
 #' @export
 
 futureAreaEstimate <- function(A.t1, year.t1, nYears, ARD = NA, PRD = NA, ARC = NA){
@@ -196,7 +277,7 @@ futureAreaEstimate <- function(A.t1, year.t1, nYears, ARD = NA, PRD = NA, ARC = 
 
 #' Extrapolate Estimate
 #'
-#' \code{extrapolateEstimate} uses rates of decline from getDeclineStats
+#' `extrapolateEstimate` uses rates of decline from getDeclineStats
 #' to extrapolate estimates to a given time
 #'
 #' @param A.t1 Area at time t1
@@ -219,7 +300,7 @@ futureAreaEstimate <- function(A.t1, year.t1, nYears, ARD = NA, PRD = NA, ARC = 
 #'   Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
 #'   List of Ecosystems Categories and Criteria, Version 1.0. Gland,
 #'   Switzerland: IUCN. ix + 94pp. Available at the following web site:
-#'   \url{https://iucnrle.org/}
+#'   <https://iucnrle.org/>
 #' @examples
 #' a.r1 <- 23.55
 #' a.r2 <- 15.79
@@ -255,7 +336,7 @@ extrapolateEstimate <- function(A.t1, year.t1, nYears, ARD = NA, PRD = NA, ARC =
 
 #' Sequential extrapolation estimate
 #'
-#' \code{sequentialExtrapolate} uses rates of decline from getDeclineStats and
+#' `sequentialExtrapolate` uses rates of decline from getDeclineStats and
 #' generates a sequence of estimates at regular time-steps. Useful for
 #' generating a sequence for plotting graphs.
 #'
@@ -273,7 +354,7 @@ extrapolateEstimate <- function(A.t1, year.t1, nYears, ARD = NA, PRD = NA, ARC =
 #'   Rodriguez, J.P. (eds.) 2016. Guidelines for the application of IUCN Red
 #'   List of Ecosystems Categories and Criteria, Version 1.0. Gland,
 #'   Switzerland: IUCN. ix + 94pp. Available at the following web site:
-#'   \url{https://iucnrle.org/}
+#'   <https://iucnrle.org/>
 #' @examples
 #' a.r1 <- 23.55
 #' a.r2 <- 15.79